pytorch_lightning 버전이 1.3.4 → 2.5.0으로 업그레이드되면서 모델 저장 파일 형태가'.ckpt'지원하지 않으므로 '.bak'로 수정함 Apply_Trainer.py - 177줄 ch_bias 필터적용 오류로 인한 코드 수정 - 361줄 pytorch_lightning 업그레이드로 인한 weight 변수 입력할때 clone().detach()사용하도록 수정 filter_back_end.py - 372줄 numpy 버전업으로 인한 dtype=np.int32 수정 custom_net.py - functional이 더 이상 pytorch_lightning.metrics에서 지원하지 않으므로 torchmetrics으로 수정

+++ filter_back_end.py
@@ -369,7 +369,7 @@ class Backgroud_remove:
         for bg_channel in range(0, gpr_aligned.shape[0]):
             for bg_depth2 in range(0, gpr_aligned.shape[1]):
                 gpr_AB[bg_channel, bg_depth2] = gpr_aligned[bg_channel, bg_depth2] \
-                                                - (np.mean(gpr_aligned[bg_channel, bg_depth2], dtype=np.int) * self.percent)
+                                                - (np.mean(gpr_aligned[bg_channel, bg_depth2], dtype=np.int32) * self.percent)
 
         return np.int16(gpr_AB)
